Move CaseInsensiveDict test to test_base_local_service and add tests for is_lambda_error_response
